adds getPrincipal method to the dao

+++ server/src/main/java/io/littlehorse/common/model/metadatacommand/subcommand/DeletePrincipalRequestModel.java
@@ -1,8 +1,12 @@
 package io.littlehorse.common.model.metadatacommand.subcommand;
 
+import com.google.protobuf.Empty;
 import com.google.protobuf.Message;
+import io.grpc.Status;
+import io.littlehorse.common.LHConstants;
 import io.littlehorse.common.LHServerConfig;
 import io.littlehorse.common.dao.MetadataProcessorDAO;
+import io.littlehorse.common.exceptions.LHApiException;
 import io.littlehorse.common.model.ClusterLevelCommand;
 import io.littlehorse.common.model.getable.objectId.PrincipalIdModel;
 import io.littlehorse.common.model.metadatacommand.MetadataSubCommand;
@@ -36,8 +40,15 @@ public class DeletePrincipalRequestModel extends MetadataSubCommand<DeletePrinci
     }
 
     @Override
-    public Message process(MetadataProcessorDAO dao, LHServerConfig config) {
+    public Empty process(MetadataProcessorDAO dao, LHServerConfig config) {
+        if (id.equals(LHConstants.ANONYMOUS_PRINCIPAL)) {
+            throw new LHApiException(
+                    Status.INVALID_ARGUMENT,
+                    "Deleting the anonymous principal is not allowed. However, "
+                            + "you can remove its permissions via the PutPrincipal request.");
+        }
+
         dao.delete(new PrincipalIdModel(id));
-        return null;
+        return Empty.getDefaultInstance();
     }
 }
